refactor: move handler dispatcher to src/events/dispatcher
